refactor(api): init src shared LearningContentResourceNotFound
