+++ api/src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js
@@ -1,0 +1,7 @@
+class LearningContentResourceNotFound extends Error {
+  constructor() {
+    super();
+  }
+}
+
+export { LearningContentResourceNotFound };
